Deep Learning/homework3/requirements.txt
matplotlib>=3.5.0
Pillow>=10.0.0
torch>=2.3.0
torchvision>=0.18.0
tensorboard>=2.0.0
termcolor==2.4.0
opencv-python>=4.10.0
